% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_afsc_species_survey.R
\name{get_afsc_species_survey}
\alias{get_afsc_species_survey}
\title{Pull taxa information for AFSC data}
\usage{
get_afsc_species_survey(spp_list)
}
\arguments{
\item{spp_list}{A list of species common_names.}
}
\description{
Pull taxa information for AFSC data
}
\author{
Chantel Wetzel
}
