% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwfsc_strata.R
\name{get_nwfsc_strata}
\alias{get_nwfsc_strata}
\title{Define NWFSC strata to use by species}
\usage{
get_nwfsc_strata(common_name)
}
\arguments{
\item{common_name}{The species common name as used in the NWFSC data warehouse}
}
\description{
Define NWFSC strata to use by species
}
\author{
Chantel Wetzel
}
