% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa.R
\name{get_taxa}
\alias{get_taxa}
\title{Pull taxa information for AFSC data}
\usage{
get_taxa(spp_list)
}
\arguments{
\item{spp_list}{A list of species common_names.}
}
\description{
Pull taxa information for AFSC data
}
\author{
Chantel Wetzel
}
