% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwfsc_strata_by_species.R
\name{nwfsc_strata_by_species}
\alias{nwfsc_strata_by_species}
\title{Define NWFSC strata to use by species}
\usage{
nwfsc_strata_by_species()
}
\description{
Define NWFSC strata to use by species
}
\author{
Chantel Wetzel
}
