% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biology.R
\name{plot_biology}
\alias{plot_biology}
\title{Plot length-weight and length-at-age}
\usage{
plot_biology(data, parameters, dir = NULL, plot = 1:2, height = 7, width = 7)
}
\arguments{
\item{data}{Dataframe of specimens data filtered down to an unique species.}

\item{parameters}{List of biological parameter estimates.}

\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{plot}{A vector of integers to specify which plots you would like. The
default is to print or save both figures, i.e., \code{plot = 1:2}. Integers
correspond to the following figures:
\enumerate{
\item Length-weight.
\item Length-at-age.
}}

\item{height, width}{Numeric values for the figure width and height in
inches. The defaults are 7 by 7 inches.}
}
\description{
Plot length-weight and length-at-age
}
\author{
Chantel Wetzel
}
