% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_design_based.R
\name{plot_design_based}
\alias{plot_design_based}
\title{Plot design-based indices}
\usage{
plot_design_based(data, dir = NULL, plot = 1, height = 7, width = 7)
}
\arguments{
\item{data}{Dataframe of design-based indices.}

\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{plot}{A vector of integers to specify which plots you would like. The
default is to print or save both figures, i.e., \code{plot = 1:2}. Integers
correspond to the following figures:
\enumerate{
\item Design based index by year and region.
\item Standardized design based index by year and region.
}}

\item{height, width}{Numeric values for the figure width and height in
inches. The defaults are 7 by 7 inches.}
}
\description{
Plot design-based indices
}
\author{
Chantel Wetzel
}
