% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_fomat_gap_biomass.R
\name{pull_format_gap_biomass}
\alias{pull_format_gap_biomass}
\title{Pull biomass estimates from AFSC surveys}
\usage{
pull_format_gap_biomass(
  survey_definition_id,
  area_id,
  taxa,
  start_year,
  end_year
)
}
\arguments{
\item{area_id}{Survey area identifier.}

\item{taxa}{Vector of information on species created by
akfingapdata::get_gap_taxonomic_classification().}

\item{start_year}{The start year of data to pull.}

\item{end_year}{The end year of data to pull.}

\item{survey_defition_id}{Survey definition identifier.}
}
\value{
Matrix of design-based biomass estimate for a species.
}
\description{
Pull biomass estimates from AFSC surveys
}
\author{
Chantel Wetzel
}
