% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_format_gap_size_comp.R
\name{pull_format_gap_size_comp}
\alias{pull_format_gap_size_comp}
\title{Pull length composition estimates from AFSC surveys}
\usage{
pull_format_gap_size_comp(
  survey_definition_id,
  area_id,
  taxa,
  start_year,
  end_year,
  comp_type = "length_cm"
)
}
\arguments{
\item{area_id}{Survey area identifier.}

\item{taxa}{Vector of information on species created by
akfingapdata::get_gap_taxonomic_classification().}

\item{start_year}{The start year of data to pull.}

\item{end_year}{The end year of data to pull.}

\item{comp_type}{Text string indicating whether to process length or age
composition data. Default "length_cm".}

\item{survey_defition_id}{Survey definition identifier.}
}
\value{
Matrix of composition estimates for a species.
}
\description{
Pull length composition estimates from AFSC surveys
}
\author{
Chantel Wetzel
}
