% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_format_nwfsc_biomass.R
\name{pull_format_nwfsc_biomass}
\alias{pull_format_nwfsc_biomass}
\title{Pull biomass estimates from NWFSC survey}
\usage{
pull_format_nwfsc_biomass(common_name, survey = "NWFSC.Combo")
}
\arguments{
\item{common_name}{The common name used in the NWFSC data warehouse.}

\item{survey_name}{The survey name used in the NWFSC data warehouse.}
}
\value{
Matrix of design-based biomass estimate for a species.
}
\description{
Pull biomass estimates from NWFSC survey
}
\author{
Chantel Wetzel
}
