% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_format_nwfsc_size_comp.R
\name{pull_format_nwfsc_size_comp}
\alias{pull_format_nwfsc_size_comp}
\title{Pull expanded length composition data for the U.S. West Coast}
\usage{
pull_format_nwfsc_size_comp(
  common_name,
  survey = "NWFSC.Combo",
  comp_type = "length_cm"
)
}
\arguments{
\item{common_name}{The common name used in the NWFSC data warehouse.}

\item{survey}{The survey name used in the NWFSC data warehouse.}

\item{comp_type}{Text string indicating whether to process length or age
composition data. Default "length_cm".}
}
\value{
Matrix of length composition estimate for a species.
}
\description{
Pull expanded length composition data for the U.S. West Coast
}
\author{
Chantel Wetzel
}
